Only add excludes to find command if it's defined
If excludes is not set in the yaml, or set to "", then the following
error occurs:

/entrypoint.sh: line 44: local: `': not a valid identifier

When excludes is empty, it causes handling of the for loop over its
contents to fail. To prevent this, check that the variable is non-empty
prior to looping over its contents.

In addition to this fix, document this feature in the readme and fix
some typos there.

Fixes #7